# swagger/portal-api.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: wicked.haufe.io Portal API
  description: The API which the API Portal uses to store and retrieve information
 
# Under securityDefinitions we declare which security schemes can be used.
# These definitions do not affect to the operations, but can be applied
# using the 'security' key at the global level or the operation level.
# In this sample, we'll see how to apply the security definition globally,
# and override it at the operation level.
#
# Note that for Access Code Flow, you must provide the Token URL.
securityDefinitions:
  # For personalized access to data
  accessCode:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/auth/local/portal-api/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/auth/local/portal-api/oauth2/authorize
    flow: accessCode
    scopes:
      read_plans: Read plans
      read_groups: Read user group definitions
      read_content: Read custom/markdown/HTML content
      health: Read out health information
      read_templates: Read templates
      read_auth_servers: Read authorization servers
      read_apis: Read API definitions
      read_users: Read user information
      write_users: Write user information
      read_applications: Read application data
      write_applications: Write application data
      read_subscriptions: Read subscriptions
      write_subscriptions: Write/edit/delete subscriptions
      read_approvals: Read pending approvals
      read_verifications: Read email and password verifications
      write_verifications: Create or delete verifications
      read_registrations: Read user registrations
      write_registrations: Write user registrations
      read_grants: Read user API grants
      write_grants: Write user API grants
      webhooks: Access and change webhooks
  
  implicit:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/portal-api/v1/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/portal-auth/api/portal-api
    flow: accessCode
    scopes:
      read_plans: Read plans
...
produces:
  - application/hal+json
host: 'localhost:3001'
basePath: /portal-api/v1
...
paths:
...
  '/apis':
    get:
      security:
      - accessCode: ['read_apis']
      - implicit: ['read_apis']
      - password: ['read_apis']
      - clientCredentials: []
      tags:
        - apis
      description: Gets the list of APIs of this API Portal.
      responses:
        '200':
          description: Successful response
          schema:
            type: object
            properties:
              apis:
                title: ArrayOfApis
                type: array
                items:
                  $ref: '#/definitions/apiType'
              _links:
                type: object
                properties:
                  self:
                    $ref: '#/definitions/hrefType'
                  plans:
                    $ref: '#/definitions/hrefType'
...
  '/content/*':
    get:
      security:
      - accessCode: ['read_content']
      - implicit: ['read_content']
      - password: ['read_content']
      - clientCredentials: ['read_content']
      tags:
        - misc
      description: Returns data from the static content section of the repository.
      produces:
        - text/markdown
        - image/jpeg
        - image/png
        - image/gif
      responses:
        '200':
          description: Successful request
        '403':
          description: Not allowed
        '404':
          description: Not found
...
  '/templates/chatbot':
    get:
      security:
      - accessCode: ['read_templates']
      - implicit: ['read_templates']
      - password: ['read_templates']
      description: Gets the Chatbot Mustache templates
      tags:
        - misc
      responses:
        200:
          description: Successful response
          schema:
            type: object
            properties:
              userLoggedIn:
                type: string
              userSignedUp:
                type: string
              userValidatedEmail:
                type: string
              applicationAdded:
                type: string
              applicationDeleted:
                type: string
              subscriptionAdded:
                type: string
              subscriptionDeleted: 
                type: string
              approvalRequired:
                type: string
              lostPasswordRequest: 
                type: string
              verifyEmailRequest: 
                type: string
  '/templates/email/{templateId}':
    get:
      security:
      - accessCode: ['read_templates']
      - implicit: ['read_templates']
      - password: ['read_templates']
      description: Get the Mustache template for a certain type of email
      tags:
        - misc
      parameters:
        - name: templateId
          in: path
          type: string
          enum:
            - "lost_password"
            - "pending_approval"
            - "verify_email"
          required: true
      responses:
        200:
          description: Successful response
        404:
          description: Not found
...
definitions:
  apiType:
    title: Api
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      desc:
        type: string
      plans:
        type: array
        items:
          type: string
      requiredGroup:
        description: 'Specify to restrict access to API, leave out to keep public'
        type: string
      tags:
        description: 'Array of tags for this API, use as "Category"'
        type: array
        items:
          type: string
      authServer:
        description: The authorization server for this API (this is optional and only applies to OAuth2 flows other than the client credentials flow)
        type: string
...
  hrefType:
    type: object
    properties:
      href:
        type: string
...